Add a target to run all MLR models
Adds a workflow target that only runs through the step of preparing data
and running the MLR models for each build name. This target or something
like it would be our primary target for initially automating the
workflow with GitHub Actions. This target does not need a custom Docker
image to run (unlike some of the later plotting rules), but it runs with
the Nextstrain base Docker image.