Add support for uploading results to S3
Adds initial code and config to upload model JSONs to S3 when the user
provides an optional config file that defines a `s3_dst` parameter. When
this is defined, the workflow will only run through the MLR model step
and then upload dated and undated model JSONs to S3. Without this
optional config file, the workflow will run as usual through the
plotting of frequencies and GAs.