Set max analysis date to 14 days prior
With a max date of 1D, the MLR model's aggregation frequency logic can
pick a final timepoint that is in the future (e.g., 2025-03-16 for
today's analysis) which doesn't make sense. The correct solution to this
problem would be to rewrite the aggregation logic to build the
timepoints back in time from the max date chosen for this analysis, but
that will require some additional work. In the mean time, we rarely get
sequences within 14 days of their collection, so this limit seems
reasonable.