Prototype haplotype assignment with Nextclade
Adds logic to assign haplotypes to all records prior to model fitting
based on Nextclade annotations and external haplotype definitions.

Haplotype definitions and colors live in the seasonal-flu repository and
get downloaded from that repository into this workflow just like we
download the Auspice config JSONs into other workflows to get clade
colors.

This commit adds a copy of the script to assign haplotypes to records in
a metadata file that contains Nextclade annotations for clades and
substitutions based on a TSV of haplotypes defined in the same format as
"augur clades" input. Ultimately, this script should become a new
subcommand in augur like `augur haplotypes`, so we can easily reuse its
functionality across all of our workflows.