[viz] add tabs to separate different models
Closes #11

Uses a very basic set of tabs to function as a table of contents,
where each tab represents a different underlying model JSON. URL queries
allow selection of the tab, which will work well with GitHub pages

Note that graph state (e.g. logit transforms) are not carried over
while swapping tabs, and aren't restored when swapping back to a
previously interacted-with tab.