[viz] allow datestamped model JSONs
This relies on the filename patterns which currently exist within the
underlying S3 bucket and is currently only accessible via a URL query.
E.g. `&date=2025-03-10` is (as of 2025-05-07) valid

Closes #16

Note that to use the snapshots functionality of nextstrain.org, as
mentioned in #16, would require the nextstrin.org resource indexer to
crawl non-auspice assets which it's not currently set up to do (but
should one day happen)