Scale generation times by aggregation frequencies
Fixes a bug in the calculation of growth advantages (GAs) when the
aggregation frequency of the count data is greater than 1 day. GAs
depend on the value of tau which is the generation time between
infections in days. However, the beta parameters from the model reflect
different units of time depending on the aggregation frequency. When
that frequency is increased, for example, to 14 days, then the beta
parameters reflect a change in frequency per 14 days instead of per 1
day. This change means that we need to adjust tau values to the same
scale as the units of the model input data. This commit updates the tau
values for each subtype, documents the original values, and updates the
aggregation frequency to units of days instead of weeks. In addition to
simplifying the configuration by using days as a consistent unit, the
original "2W-SUN" aggregation could also produce time bins where the
last bin represented a future date (i.e., the next Sunday up to 2 weeks
into the future).