Estimate frequencies from a given max date
Adds a max date argument to the MLR model script which allows us to
define date intervals to estimate frequencies by starting at that max
date and working backward in steps defined by the aggregation frequency.
In contrast, the original aggregation frequency method started defining
intervals based on the earliest date in the input data and stepping
forward in time from that early date. The resulting intervals from that
approach are rarely ever what we want in practice. The change in this
commit makes the date aggregation intervals for the MLR model more
consistent with the approach used in augur frequencies.