Use 6 forecast steps of 14 days each
After running the models with the latest evofr, I needed to set the
forecast horizon to the number of steps into the future I want to match
the number of total days to be forecasted. The last forecast date is L
steps times the aggregation frequency into the future. With an
aggregation frequency (step size) of 14 days, I've set the forecast
horizon to 6 steps to get 84-day forecasts.