Display multiple variant classifications
Update forecasts viz interface to display MLR results for multiple
variant classifications on a single page per subtype and geographic
resolution, following the pattern from forecasts-ncov [1].

Since the `useModelData` functions lives outside of this repo and
expects a single object with a `modelUrl` attribute, this commit creates
two separate config objects with one per variant classification. We call
the model URL function once per config after the initial config object
is copied for each variant classification. This function now accepts
arguments for the variant classification (to get the correct S3 URL) and
for model date (so we only need the date-based update logic in one
place).

As I implemented this expanded version of the display, I found it
simpler to write the panel titles (the `h2` tags) and descriptions right
in the HTML instead of maintaining a separate title per variant
classification, subtype, and geographic resolution. This new
implementation mimics the forecasts-ncov at the expense of some
flexibility in defining different titles and descriptions per
subtype (which we have never needed to do).

[1] https://github.com/nextstrain/forecasts-ncov/blob/940791b/viz/src/App.jsx

Co-authored-by: james hadfield <hadfield.james@gmail.com>